Moved flavors and dimensions creation to `AndroidBuildGradleBlueprint`
